lk/requirements.txt
django==2.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
clearbit==0.1.7
requests==2.19.1